+++ example/lib/shared/const/store.dart
@@ -63,6 +63,11 @@ class Store {
   // Default value for the deviceSize, also used to reset settings.
   static const double defaultDeviceSize = 200;
 
+  // Key used to read and save the showSchemeInput value.
+  static const String keyShowSchemeInput = 'showSchemeInput';
+  // Default value for the showSchemeInput, also used to reset settings.
+  static const bool defaultShowSchemeInput = false;
+
   // Key used to read and save the useTextTheme value.
   static const String keyUseTextTheme = 'useTextTheme';
   // Default value for the useSubThemes, also used to reset settings.