+++ CHANGELOG.md
@@ -26,13 +26,16 @@ All notable changes to the **FlexColorScheme** (FCS) package are documented here
 - Added "set to M3 default values" button to NavigationBar and NavigationRail settings.
 - Added a feature to select which panel is shown in page view as the second panel. It is no longer locked to the code-view panel, it can be any of the available panels, enabling studying chosen panels side by side.
 - Added an example themed-app panel view with a few screens. Can be used as a side screen to show what an app using the theme might look like. Screens still need demo content.
+- Hides scheme input color definitions by default. You can turn it back ON if so desired. Showing input scheme color values shows the used color values before any input modifiers are used on them. Like, seed generation, swapping legacy colors, swapping primary and secondary colors, seed generation, input color limiters, computed dark theme and using M3 error colors on legacy M2 schemes. In older versions of FCS, the scheme input color values were always shown. This may be confusing and they are now hidden by default. Showing them may be useful to understand how the scheme defined input colors are being modified by input modifiers, but it is cluttered and confusing. If ON, scheme input color values show the colors before input modifiers, and the surrounding color shows the effective theme's ColorScheme. 
 
 **TODO BEFORE FCS STABLE 7.0 RELEASE**
 
+- Add TextFiled Demo setting example(s).
+- Add TextField prefixIcon color, M2 default to base, M3 Default to onSurface.
 - Add BottomAppBarColor background color theming.
 - Add NavigationDrawer width and selected item colors. 
-- Add some theming options for the new MenuBar theme.  
-- Playground: Hide input color. Make it default to hide them.  
+- Maybe: Add some theming options for the new MenuBar theme.  
+  
 - MUST DO: Review and do actionable TODOs in the code.
 - MUST DO: Review and fix test coverage. Test new colors too.
 - MUST DO: Review and test all tutorial examples.